requirements.txt
asgiref
Django==4.0.4
Pillow~=9.0.1
pytz~=2019.3
sqlparse
django-template-maths
django-htmx
django-widget-tweaks
djangorestframework
markdown
django-filter
serializers